stock_trading_app_complete/stock_trading_app_complete/requirements.txt
flask
flask_sqlalchemy
flask_login
python-dotenv
requests
pytest
gunicorn
pytest-mock